move hasRootToLoopLinearTransformations
